+++ cmd/cremengine/config/ChangeLog.md
@@ -1,7 +1,9 @@
 # Change Log
 
 ## Version 0.3 (TBD):
+### Bugs Fixed
 * Fixed bug in Subcatchment parser, stopping WetlandsEstablishment action from having state set.
+* Fixed bug in Subcatchment parser, where semantically invalid action state changes would respond with success. 
 
 ## Version 0.2 (02 March 2021):
 ### New Features